made changes as per the comments
