+++ server/gql/subscriptions/purchasedProductSubscription/purchasedProductSubsUtil.js
@@ -1,9 +1,3 @@
-import { pubsub } from '@server/utils/pubsub';
-
-export function getAsyncInterator(topicName) {
-  return () => pubsub.asyncIterator(topicName);
-}
-
-export function getFilteredSubscription(payload, variables) {
-  return Number(payload.notifications.storeId) === variables.storeId;
+export function checkFilterCondition(payload, variables) {
+  return payload.notifications.storeId === variables.storeId;
 }
