+++ server/gql/subscriptions/purchasedProductSubscription/tests/index.test.js
@@ -1,6 +1,6 @@
 import { getResponse } from '@utils/testUtils';
+import { checkFilterCondition } from '../purchasedProductSubsUtil';
 import * as module from 'graphql-subscriptions';
-import '@utils/pubsub';
 
 describe('Subscription tests', () => {
   it('should add a subscription', async () => {
@@ -16,6 +16,6 @@ describe('Subscription tests', () => {
   `;
     const spy = jest.spyOn(module, 'withFilter');
     await getResponse(subscription);
-    expect(spy).toBeCalledTimes(1);
+    expect(spy).toBeCalledWith(expect.any(Function), checkFilterCondition);
   });
 });
