+++ server/gql/subscriptions/purchasedProductSubscription/tests/purchasedProductSubsUtil.test.js
@@ -1,15 +1,6 @@
-import { SUBSCRIPTION_TOPICS } from '@server/utils/constants';
-import { pubsub } from '@server/utils/pubsub';
-import { getAsyncInterator, getFilteredSubscription } from '../purchasedProductSubsUtil';
+import { checkFilterCondition } from '../purchasedProductSubsUtil';
 
 describe('Purchased Products subscription utils test', () => {
-  describe('getAsyncInterator tests', () => {
-    it('should call the async iterator with the given topic', () => {
-      const spy = jest.spyOn(pubsub, 'asyncIterator');
-      getAsyncInterator(SUBSCRIPTION_TOPICS.NOTIFICATIONS)();
-      expect(spy).toBeCalled();
-    });
-  });
   describe('getFilteredSubscription tests', () => {
     it('should return true if the supplierId in payload is equal to supplierId in variables ', () => {
       const payload = {
@@ -20,7 +11,7 @@ describe('Purchased Products subscription utils test', () => {
       const variables = {
         storeId: 1
       };
-      const res = getFilteredSubscription(payload, variables);
+      const res = checkFilterCondition(payload, variables);
       expect(res).toBeTruthy();
     });
   });
