+++ server/utils/iterator.js
@@ -1,0 +1,4 @@
+import { pubsub } from '@utils/pubsub';
+export function getAsyncInterator(topicName) {
+  return () => pubsub.asyncIterator(topicName);
+}
